Fix code defects generated by coverity tool.
CID:61142
Invalid Assignment: The type size of both side variables are different.
"wBuffIdx" is 4 (unsigned int) and "wBuffIdx +2UL" is
8 (unsigned long)

+++ SN100x/halimpl/dnld/phDnldNfc_Internal.cc
@@ -762,7 +762,7 @@ static NFCSTATUS phDnldNfc_CreateFramePld(pphDnldNfc_DlContext_t pDlContext) {
              &(pDlContext->tRWInfo.wBytesToSendRecv),
              (sizeof(pDlContext->tRWInfo.wBytesToSendRecv)));
 
-      wBuffIdx += sizeof(pDlContext->tRWInfo.wBytesToSendRecv);
+      wBuffIdx += (uint32_t)sizeof(pDlContext->tRWInfo.wBytesToSendRecv);
 
       memcpy(&(pDlContext->tCmdRspFrameInfo.aFrameBuff[wBuffIdx]),
              &(pDlContext->tRWInfo.dwAddr), sizeof(pDlContext->tRWInfo.dwAddr));
